Move events listener to dedicated modules
