+++ src/popup/components/list/assets/styles/list.css
@@ -98,6 +98,7 @@
 		cursor: pointer;
 		position: relative;
 		text-decoration: none;
+		text-align: left;
 
 		* {
 			pointer-events: none;
